backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.2
python-multipart==0.0.17
httpx[socks]==0.28.1  # SOCKS代理支持（国际词典API需要）
ebooklib==0.18
beautifulsoup4==4.12.3
aiofiles==24.1.0
nltk==3.9.1
python-dotenv==1.0.0
oss2==2.18.4
